pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "llmbrix"
authors = [
    { name = "Matej Kvassay", email = "matejkvassay5@gmail.com" }
]
description = "Low abstraction agentic LLM framework."
version = "2.0.0a1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz",
    "pillow",
    "google-genai>=1.56.0",
    "pydantic",
    "sympy",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha", # Consider 4 - Beta or 5 - Production/Stable as it matures
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries"
]
keywords = ["LLM", "agent", "framework", "AI", "chatbot", 'llmbrix', 'llmbricks']

[project.urls]
"Homepage" = "https://github.com/matejkvassay/LLMBrix"
"Repository" = "https://github.com/matejkvassay/LLMBrix"
"Issues" = "https://github.com/matejkvassay/LLMBrix/issues"
"Documentation" = "https://github.com/matejkvassay/LLMBrix#readme"

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "build",
    "twine"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["llmbrix"]
